handle import better use argument for fake or real

+++ backend/convert/import.sh
@@ -4,43 +4,70 @@ SCRIPT_PATH=$(readlink -f $BASEDIR)
 
 CURRENTSEASON=2014
 echo CURRENTSEASON=$CURRENTSEASON
+arg=$1
 
-DBCMD="mysql -u roprotokol -proprotokol roprotokol"
+if [ "x"$arg == "x" ]
+then
+   echo usage:
+   echo   import.sh fake
+   echo   import.sh real
+   echo fake use DSR data for boats, destinations, etc but generate fake informations about rowers
+   echo real is for importing real DSR data, it requires that you have SQL dumps from the database
+   exit 0
+fi
+DBCMD="mysql -u roprotokol roprotokol"
 
-for tb in Båd Bådindstilling BådKategori Fejl_system Fejl_tblMembersSportData Fejl_tur Gruppe  Kajak_typer Kommentar LåsteBåde Medlem Motionstatus Postnr Reservation Skade TurDeltager TurType  Vintervedligehold Destination Kajak_anvendelser Tur tblMembersSportData; do
+#if you you a password, put DBCMD="mysql -u roprotokol -p password roprotokol" in secret.sh
+. $SCRIPT_PATH/secret.sh
+
+for tb in Båd Bådindstilling BådKategori Gruppe  Kajak_typer Kommentar LåsteBåde Medlem Motionstatus Postnr Reservation Skade TurType Destination Kajak_anvendelser; do
     echo DO IMPORT $tb
     echo
     $DBCMD -e "TRUNCATE TABLE $tb;"
     $DBCMD < $SCRIPT_PATH/data/$tb.sql
 done
 
-for SEASON in $(seq 2010 2013); do
-    echo SEASON $SEASON; 
-    for ST in Tur Turdeltager; do
-	tb=${ST}_backup${SEASON}
-	echo make table $tb
-	$DBCMD -e "DROP TABLE IF EXISTS $tb" 
-	$DBCMD -e "CREATE TABLE $tb AS SELECT * from ${ST/deltager/Deltager} WHERE 1=2"
+if [ $arg=="fake" ]; then
+    $SCRIPT_PATH/../tests/fakedata.py
+elif [ $arg=="real" ]; then
+    for tb in Fejl_tblMembersSportData Fejl_system Fejl_tur TurDeltager Vintervedligehold Tur tblMembersSportData; do
+	echo DO IMPORT $tb
+	echo
+	$DBCMD -e "TRUNCATE TABLE $tb;"
 	$DBCMD < $SCRIPT_PATH/data/$tb.sql
-    done
 done
 
-for SEASON in $(seq 2010 2013); do
-    $DBCMD -e "INSERT INTO Trip (TripID,Season,BoatID,OutTime,InTime,ExpectedIn,Destination,Meter,TripTypeID,Comment,CreatedDate,EditDate,Initials,DESTID) \
+
+    for SEASON in $(seq 2010 2013); do
+	echo SEASON $SEASON; 
+	for ST in Tur Turdeltager; do
+	    tb=${ST}_backup${SEASON}
+	    echo make table $tb
+	    $DBCMD -e "DROP TABLE IF EXISTS $tb" 
+	    $DBCMD -e "CREATE TABLE $tb AS SELECT * from ${ST/deltager/Deltager} WHERE 1=2"
+	    $DBCMD < $SCRIPT_PATH/data/$tb.sql
+	done
+    done
+
+    for SEASON in $(seq 2010 2013); do
+	$DBCMD -e "INSERT INTO Trip (TripID,Season,BoatID,OutTime,InTime,ExpectedIn,Destination,Meter,TripTypeID,Comment,CreatedDate,EditDate,Initials,DESTID) \
      SELECT TurID,${SEASON},FK_BådID,Ud,Ind,ForvInd,Destination,Meter,FK_TurTypeID,Kommentar,OprettetDato,RedigeretDato,Initialer,DESTID FROM Tur_backup${SEASON}"
 
-    $DBCMD -e "INSERT INTO TripMember (TripID, Season, Seat, MemberID,MemberName,CreatedDate,EditDate,Initials) \
-    SELECT FK_TurID, ${SEASON}, Plads, FK_MedlemID,Navn,OprettetDato,RedigeretDato,Initialer FROM Turdeltager_backup${SEASON}"
-    $DBCMD -e "DROP TABLE Tur_backup${SEASON}"
-    $DBCMD -e "DROP TABLE Turdeltager_backup${SEASON}"
+	$DBCMD -e "INSERT INTO TripMember (TripID, Season, Seat, MemberID,MemberName,CreatedDate,EditDate,Initials) \
+        SELECT FK_TurID, ${SEASON}, Plads, FK_MedlemID,Navn,OprettetDato,RedigeretDato,Initialer FROM Turdeltager_backup${SEASON}"
+	$DBCMD -e "DROP TABLE Tur_backup${SEASON}"
+	$DBCMD -e "DROP TABLE Turdeltager_backup${SEASON}"
 done
 
-SEASON=$CURRENTSEASON
-$DBCMD -e "INSERT INTO Trip (TripID,Season,BoatID,OutTime,InTime,ExpectedIn,Destination,Meter,TripTypeID,Comment,CreatedDate,EditDate,Initials,DESTID) \
+    SEASON=$CURRENTSEASON
+    $DBCMD -e "INSERT INTO Trip (TripID,Season,BoatID,OutTime,InTime,ExpectedIn,Destination,Meter,TripTypeID,Comment,CreatedDate,EditDate,Initials,DESTID) \
      SELECT TurID,${SEASON},FK_BådID,Ud,Ind,ForvInd,Destination,Meter,FK_TurTypeID,Kommentar,OprettetDato,RedigeretDato,Initialer,DESTID FROM Tur"
 
-$DBCMD -e "INSERT INTO TripMember (TripID, Season,Seat, MemberID,MemberName,CreatedDate,EditDate,Initials) \
+    $DBCMD -e "INSERT INTO TripMember (TripID, Season,Seat, MemberID,MemberName,CreatedDate,EditDate,Initials) \
     SELECT   FK_TurID, ${SEASON}, Plads, FK_MedlemID,Navn,OprettetDato,RedigeretDato,Initialer FROM TurDeltager"
-$DBCMD -e "DROP TABLE Tur"
-$DBCMD -e "DROP TABLE TurDeltager"
-$DBCMD < $SCRIPT_PATH/konvertRights.sql
+    $DBCMD -e "DROP TABLE Tur"
+    $DBCMD -e "DROP TABLE TurDeltager"
+    $DBCMD < $SCRIPT_PATH/konvertRights.sql
+else
+    echo unknown argument
+fi
